requirements.txt
requests
python-telegram-bot
sentence-transformers
faiss-cpu
rapidfuzz
torch
numpy
python-dotenv